requirements.txt
pytest==6.2.4
allure-pytest==2.13.5

Appium-Python-Client~=4.2.0
selenium~=4.25.0
PyYAML~=6.0.2